refactor: Update navigation bar to use NavigationItems component for larger screens
