+++ app/(sub-page)/components/navigation-bar.tsx
@@ -5,7 +5,7 @@ import SideNavigationBar from './side-navigation-bar';
 import UserInfoNavigator from '@/app/ui/user/user-info-navigator/user-info-navigator';
 import SignButtonGroup from '@/app/ui/user/user-info-navigator/sign-button-group';
 import Link from 'next/link';
-import Button from '@/app/ui/view/atom/button/button';
+import NavigationItems from './navigation-items';
 
 export default function NavigationBar() {
   return (
@@ -17,25 +17,8 @@ export default function NavigationBar() {
         <SideNavigationBar header={<UserInfoNavigator />} content={<div>콘텐츠</div>} footer={<SignButtonGroup />} />
       </Responsive>
       <Responsive minWidth={1024}>
-        <div className="flex">
-          <NavigationItem href={'/sign-in'} label="로그인" />
-          <NavigationItem href={'/tutorial'} label="튜토리얼" />
-          <NavigationItem href={'/'} label="팀소개" />
-        </div>
+        <NavigationItems />
       </Responsive>
     </div>
-  );
-}
-
-type NavigationItemProps = {
-  href: string;
-  label: string;
-};
-
-function NavigationItem({ href, label }: NavigationItemProps) {
-  return (
-    <Link href={href}>
-      <Button size={'xs'} className="text-white hover:text-slate-400" variant={'text'} label={label} />
-    </Link>
   );
 }
